Swedish: rename pascal case to money case
In pascal case, all words are changed so that the
begin with an upper case, and everything else is
left lower case, than all spaces are removes.

Swedish does not have spaces in spelled out numbers
so calling it pascal case is in error, it's just
the components (not all though) that are changed to
begin with an upper case letter. Since this case
is basically only and always used when spelling
out large sums of money, the name money case is a
much better description.

Signed-off-by: Mattias Andrée <maandree@kth.se>